Vectorize over inputs
These changes update the code to vectorize over multiple inputs guesses and shortcuts the integration phase after *all* contours have hit an endpoint or have closed.

Note: as a result each path is integrated for as long as the longest contour, while not as efficient as it could be, it is far cleaner than the code before was.

Once the integration finishes in each direction the results are masked to the point where the contour first closes/stops.  All other values are written to NaN.  The reverse path is inverted and placed before the forward path (with the initial point in the middle).  After stacking the entire vectored is `roll`ed so the non-NaN values are at the front.

The fixed output size means this now plays nicely with `jit`.

A helper function `path_reduce` is provided to mask out the NaN values (and as a result splits the output into a list of arrays).