Update code to allow auto-diff
These changes included a full refactor of the code into a class object, greatly improving the readability of many of the functions.

The other major change is the inclusion of `jax.lax.custom_root` for the Newton's step, this allows both forward and reverse auto-diff to work as expected and be much faster to compute.  This enables contours to be included in loss functions if needed.